refactor: 탭 파일 이동 및 게시글 지역란 수정
